page_loadsample: use INT_SHAPED_PTR and PTR_SHAPED_INT
also changed the definition of PTR_SHAPED_INT to cast to intptr_t
first, since casting from integer of different size screws up the
signed bits (no sign extension)

+++ include/headers.h
@@ -292,7 +292,7 @@ struct stat {
 #endif
 
 #define INT_SHAPED_PTR(v) ((intptr_t)(void*)(v))
-#define PTR_SHAPED_INT(i) ((void*)(i))
+#define PTR_SHAPED_INT(i) ((void*)(intptr_t)(i))
 
 /* ------------------------------------------------------------------------ */
